conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.5.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.7.15=h06a4308_0
cuda-cccl_linux-64=12.4.127=h06a4308_2
cuda-cudart=12.1.105=0
cuda-cudart-dev_linux-64=12.4.127=hd681fbe_0
cuda-cudart-static_linux-64=12.4.127=hd681fbe_0
cuda-cudart_linux-64=12.4.127=hd681fbe_0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.127=0
cuda-runtime=12.1.0=0
cuda-version=12.4=hbda6634_3
cudensitymat=0.2.0=nompi_h6099206_100
cupy=13.4.1=py313hdf5e20e_0
cupy-core=13.4.1=py313h2626f57_0
cuquantum-python=25.06.0=py313ha320421_0
custatevec=1.9.0=hcd2ec93_0
cutensor=2.2.0.0=h2d75abc_1
cutensornet=2.8.0=nompi_h6099206_100
deprecated=1.2.13=py313h06a4308_0
expat=2.7.1=h6a678d5_0
fastrlock=0.8.3=py313h5eee18b_0
filelock=3.17.0=py313h06a4308_0
fsspec=2025.7.0=py313h7040dfc_0
gmp=6.3.0=h6a678d5_0
gmpy2=2.2.1=py313h5eee18b_0
importlib-metadata=8.5.0=py313h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.6=py313h06a4308_0
ld_impl_linux-64=2.40=h12ee557_0
libabseil=20250127.0=cxx17_h6a678d5_0
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.9.1.3=0
libcurand=10.3.5.147=h99ab3db_1
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libffi=3.4.4=h6a678d5_1
libgcc=15.1.0=h767d61c_4
libgcc-ng=15.1.0=h69a702a_4
libgfortran=15.1.0=h69a702a_4
libgfortran5=15.1.0=hcea5267_4
libgomp=15.1.0=h767d61c_4
libmpdec=4.0.0=h5eee18b_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libopenblas=0.3.30=pthreads_h94d23a6_1
libprotobuf=5.29.3=h3cdef7c_1
libstdcxx=15.1.0=h8f9b012_4
libstdcxx-ng=11.2.0=h1234567_1
libtorch=2.6.0=cpu_mkl_h881e62d_106
libuuid=1.41.5=h5eee18b_0
libuv=1.48.0=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
markupsafe=3.0.2=py313h5eee18b_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py313h5eee18b_2
mkl_fft=1.3.11=py313h5eee18b_0
mkl_random=1.2.8=py313h06d7b56_0
mpc=1.3.1=h5eee18b_0
mpfr=4.2.1=h5eee18b_0
mpmath=1.3.0=py313h06a4308_0
ncurses=6.5=h7934f7d_0
networkx=3.4.2=py313h06a4308_0
numpy=2.3.1=py313h8d96ed3_0
numpy-base=2.3.1=py313h8e760e0_0
openssl=3.0.17=h5eee18b_0
opentelemetry-api=1.30.0=py313h06a4308_0
pip=25.1=pyhc872135_2
pthread-stubs=0.3=h0ce48e5_1
python=3.13.5=h4612cfd_100_cp313
python_abi=3.13=0_cp313
pytorch=2.6.0=cpu_mkl_py313hfd6889d_106
pytorch-cuda=12.1=ha16c6d3_6
readline=8.3=hc2a1206_0
setuptools=72.1.0=py313h06a4308_0
sleef=3.5.1=h5eee18b_2
sqlite=3.50.2=hb25bd0a_1
sympy=1.13.3=py313h06a4308_1
tbb=2021.8.0=hdb19cb5_0
tensorsb=0.1.0=pypi_0
tk=8.6.14=h993c535_1
typing_extensions=4.12.2=py313h06a4308_0
tzdata=2025b=h04d1e81_0
wheel=0.45.1=py313h06a4308_0
wrapt=1.17.0=py313h5eee18b_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
zipp=3.21.0=py313h06a4308_0
zlib=1.2.13=h5eee18b_1
